+++ lib/src/api/entities/account.dart
@@ -12,4 +12,6 @@ abstract class Account extends RestrrEntity {
   Future<bool> delete();
 
   Future<Account> update({String? name, String? description, String? iban, int? originalBalance, Id? currency});
+
+  Future<Paginated<Transaction>> retrieveAllTransactions({int page = 1, int limit = 25, bool forceRetrieve = false});
 }